Move code related to Elasticsearch to own package
