+++ georocket-server/src/main/java/io/georocket/index/IndexerVerticle.java
@@ -16,15 +16,8 @@ import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import org.apache.commons.exec.CommandLine;
-import org.apache.commons.exec.DefaultExecuteResultHandler;
-import org.apache.commons.exec.DefaultExecutor;
-import org.apache.commons.exec.ExecuteException;
-import org.apache.commons.exec.Executor;
-import org.apache.commons.exec.ShutdownHookProcessDestroyer;
-import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.io.IOUtils;
-import org.apache.commons.lang.SystemUtils;
+import org.apache.commons.lang3.tuple.Pair;
 import org.jooq.lambda.tuple.Tuple;
 import org.yaml.snakeyaml.Yaml;
 
@@ -32,6 +25,9 @@ import com.google.common.collect.ImmutableList;
 
 import io.georocket.constants.AddressConstants;
 import io.georocket.constants.ConfigConstants;
+import io.georocket.index.elasticsearch.ElasticsearchClient;
+import io.georocket.index.elasticsearch.ElasticsearchInstaller;
+import io.georocket.index.elasticsearch.ElasticsearchRunner;
 import io.georocket.index.xml.XMLIndexer;
 import io.georocket.index.xml.XMLIndexerFactory;
 import io.georocket.query.DefaultQueryCompiler;
@@ -43,9 +39,7 @@ import io.georocket.util.AsyncXMLParser;
 import io.georocket.util.MapUtils;
 import io.georocket.util.RxUtils;
 import io.georocket.util.XMLStartElement;
-import io.vertx.core.AsyncResult;
 import io.vertx.core.Future;
-import io.vertx.core.Handler;
 import io.vertx.core.Vertx;
 import io.vertx.core.impl.NoStackTraceThrowable;
 import io.vertx.core.json.JsonArray;
@@ -76,6 +70,11 @@ public class IndexerVerticle extends AbstractVerticle {
 
   protected static final String INDEX_NAME = "georocket";
   protected static final String TYPE_NAME = "object";
+  
+  /**
+   * Runs Elasticsearch
+   */
+  protected ElasticsearchRunner runner;
   
   /**
    * The Elasticsearch client
@@ -109,8 +108,9 @@ public class IndexerVerticle extends AbstractVerticle {
     indexerFactories = createIndexerFactories();
 
     createElasticsearchClient()
-      .subscribe(esclient -> {
-        client = esclient;
+      .subscribe(es -> {
+        client = es.getKey();
+        runner = es.getValue();
         
         queryCompiler = new DefaultQueryCompiler(indexerFactories);
         store = StoreFactory.createStore((Vertx)vertx.getDelegate());
@@ -141,9 +141,9 @@ public class IndexerVerticle extends AbstractVerticle {
   /**
    * Create an Elasticsearch client. Either start an Elasticsearch instance or
    * connect to an external one - depending on the configuration.
-   * @return an observable emitting the Elasticsearch client
+   * @return an observable emitting an Elasticsearch client and runner
    */
-  private Observable<ElasticsearchClient> createElasticsearchClient() {
+  private Observable<Pair<ElasticsearchClient, ElasticsearchRunner>> createElasticsearchClient() {
     JsonObject config = vertx.getOrCreateContext().config();
     boolean embedded = config.getBoolean(ConfigConstants.INDEX_ELASTICSEARCH_EMBEDDED, true);
     String host = config.getString(ConfigConstants.INDEX_ELASTICSEARCH_HOST, "localhost");
@@ -178,122 +178,23 @@ public class IndexerVerticle extends AbstractVerticle {
     
     if (!embedded) {
       // just return the client
-      return Observable.just(client);
+      return Observable.just(Pair.of(client, null));
     }
     
     return client.isRunning().flatMap(running -> {
       if (running) {
         // we don't have to start Elasticsearch again
-        return Observable.just(client);
+        return Observable.just(Pair.of(client, null));
       }
       
       // install Elasticsearch, start it and then create the client
       ElasticsearchInstaller installer = new ElasticsearchInstaller(vertx);
+      ElasticsearchRunner runner = new ElasticsearchRunner(getVertx());
       return installer.download(elasticsearchDownloadUrl, elasticsearchInstallPath)
-        .flatMap(path -> runElasticsearch(host, port, path))
-        .flatMap(v -> waitUntilElasticsearchRunning(client))
-        .map(v -> client);
+        .flatMap(path -> runner.runElasticsearch(host, port, path))
+        .flatMap(v -> runner.waitUntilElasticsearchRunning(client))
+        .map(v -> Pair.of(client, runner));
     });
-  }
-  
-  /**
-   * Run Elasticsearch
-   * @param host the host Elasticsearch should bind to
-   * @param port the port Elasticsearch should listen on
-   * @param elasticsearchInstallPath the path where Elasticsearch is installed
-   * @return an observable that emits exactly one item when Elasticsearch has started
-   */
-  private Observable<Void> runElasticsearch(String host, int port,
-      String elasticsearchInstallPath) {
-    JsonObject config = vertx.getOrCreateContext().config();
-    String storage = config.getString(ConfigConstants.STORAGE_FILE_PATH);
-    String root = storage + "/index";
-    
-    ObservableFuture<Void> observable = RxHelper.observableFuture();
-    Handler<AsyncResult<Void>> handler = observable.toHandler();
-    
-    ((Vertx)vertx.getDelegate()).<Void>executeBlocking(f -> {
-      log.info("Starting Elasticsearch ...");
-      
-      // get Elasticsearch executable
-      String executable = FilenameUtils.separatorsToSystem(
-          elasticsearchInstallPath);
-      executable = FilenameUtils.concat(executable, "bin");
-      if (SystemUtils.IS_OS_WINDOWS) {
-        executable = FilenameUtils.concat(executable, "elasticsearch.bat");
-      } else {
-        executable = FilenameUtils.concat(executable, "elasticsearch");
-      }
-      
-      // start Elasticsearch
-      CommandLine cmdl = new CommandLine(executable);
-      cmdl.addArgument("--cluster.name");
-      cmdl.addArgument("georocket-cluster");
-      cmdl.addArgument("--node.name");
-      cmdl.addArgument("georocket-node");
-      cmdl.addArgument("--network.host");
-      cmdl.addArgument(host);
-      cmdl.addArgument("--http.port");
-      cmdl.addArgument(String.valueOf(port));
-      cmdl.addArgument("--path.home");
-      cmdl.addArgument(elasticsearchInstallPath);
-      cmdl.addArgument("--path.data");
-      cmdl.addArgument(root + "/data");
-      
-      Executor executor = new DefaultExecutor();
-      executor.setProcessDestroyer(new ShutdownHookProcessDestroyer());
-      
-      try {
-        executor.execute(cmdl, new DefaultExecuteResultHandler() {
-          @Override
-          public void onProcessComplete(final int exitValue) {
-            log.info("Elasticsearch quit with exit code: " + exitValue);
-          }
-          
-          @Override
-          public void onProcessFailed(final ExecuteException e) {
-            log.error("Elasticsearch execution failed", e);
-          }
-        });
-        f.complete();
-      } catch (IOException e) {
-        f.fail(e);
-      }
-    }, handler);
-    
-    return observable;
-  }
-  
-  /**
-   * Wait 60 seconds or until Elasticsearch is up and running, whatever
-   * comes first
-   * @param client the client to use to check if Elasticsearch is running
-   * @return an observable that emits exactly one item when Elasticsearch
-   * is running
-   */
-  private Observable<Void> waitUntilElasticsearchRunning(
-      ElasticsearchClient client) {
-    Scheduler scheduler = RxHelper.scheduler(getVertx());
-    final Throwable repeat = new NoStackTraceThrowable("");
-    return Observable.<Boolean>create(subscriber -> {
-      client.isRunning().subscribe(subscriber);
-    }).flatMap(running -> {
-      if (!running) {
-        return Observable.error(repeat);
-      }
-      return Observable.just(running);
-    }).retryWhen(errors -> {
-      Observable<Throwable> o = errors.flatMap(t -> {
-        if (t == repeat) {
-          // Elasticsearch is still not up, retry
-          return Observable.just(t);
-        }
-        // forward error
-        return Observable.error(t);
-      });
-      // retry for 60 seconds
-      return RxUtils.makeRetry(60, 1000, scheduler, log).call(o);
-    }, scheduler).map(r -> null);
   }
   
   /**
