+++ georocket-server/src/main/java/io/georocket/index/elasticsearch/ElasticsearchInstaller.java
@@ -1,4 +1,4 @@
-package io.georocket.index;
+package io.georocket.index.elasticsearch;
 
 import java.io.File;
 import java.io.FileOutputStream;
