+++ georocket-server/src/test/java/io/georocket/index/elasticsearch/ElasticsearchInstallerTest.java
@@ -1,4 +1,4 @@
-package io.georocket.index;
+package io.georocket.index.elasticsearch;
 
 import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
 import static com.github.tomakehurst.wiremock.client.WireMock.get;
@@ -22,6 +22,7 @@ import com.github.tomakehurst.wiremock.http.Fault;
 import com.github.tomakehurst.wiremock.junit.WireMockRule;
 
 import io.georocket.NetUtils;
+import io.georocket.index.elasticsearch.ElasticsearchInstaller;
 import io.vertx.core.Vertx;
 import io.vertx.ext.unit.Async;
 import io.vertx.ext.unit.TestContext;
