+++ georocket-server/src/test/resources/io/georocket/index/elasticsearch/elasticsearch-dummy.zip
